// web/src/api/generated/model/updateInputBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { SignedRecord } from "./signedRecord.ts";

export interface UpdateInputBody {
  password_envelope?: SignedRecord;
  private_envelope?: SignedRecord;
  recovery_envelope?: SignedRecord;
}